u-boot-android.bb: Fix the build of the u-boot-android
The U-Boot that we are using for Android is built not from the
sources, defined in Yocto, but from a custom source code, which is
located in https://github.com/xen-troops/u-boot

After switching to the Kirkstone Yocto release, the LIC_FILES_CHKSUM,
which is specified in the 'u-boot-common.inc' stopped fitting the
license file in the xen-troops/u-boot repository on GitHub.

Issue was causing a build failure of the Dom0, when built with
ENABLE_ANDROID parameter.

This patch adds an override of the LIC_FILES_CHKSUM for u-boot-android.bb.
It does not make sense to override sources while keeping the external
reference to the license checksum. Such an approach will continuously
break.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>