domu.service: Add dependency on domd.service
After switching off the 'displbe' and 'sndbe' services, it turned out
that DomU may start before DomD. That is causing issues, as some
parameters are strictly relying on DomD having the Xen domain id equal
to '1'.

In order to achieve the correct starting order of the domains, we've
added dependency of domu.service on domd.service.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>